Create Importing comments from Zoom as main video subtitles (srt format)
When one has a Zoom meeting, it is possible to download a .mp4 video and a .txt file. This code imports the former media to create another .srt file to be loaded while playing the video later on, so comments are displayed as subtitles. Handy when trying to relate the video to the online written channel discussion.